Change rr::Emitter to use longs instead of ints.

+++ src/render/emitter.cpp
@@ -21,8 +21,8 @@ namespace rr {
 namespace {
 
 #ifdef TRACK_CAPACITY
-void log_capacity_change( bool on, int capacity_before,
-                          int capacity_after ) {
+void log_capacity_change( bool on, long capacity_before,
+                          long capacity_after ) {
   DCHECK( capacity_after >= capacity_before );
   if( on && capacity_after != capacity_before )
     fmt::print(
@@ -32,8 +32,8 @@ void log_capacity_change( bool on, int capacity_before,
         double( capacity_after ) / double( capacity_before ) );
 }
 #else
-void log_capacity_change( bool /*on*/, int /*capacity_before*/,
-                          int /*capacity_after*/ ) {}
+void log_capacity_change( bool /*on*/, long /*capacity_before*/,
+                          long /*capacity_after*/ ) {}
 #endif
 
 } // namespace
@@ -42,34 +42,34 @@ void log_capacity_change( bool /*on*/, int /*capacity_before*/,
 ** Emitter
 *****************************************************************/
 void Emitter::emit( GenericVertex const& vert ) {
-  int capacity_before = buffer_->capacity();
-  DCHECK( pos_ <= int( buffer_->size() ) );
-  if( pos_ < int( buffer_->size() ) )
+  long capacity_before = buffer_->capacity();
+  DCHECK( pos_ <= long( buffer_->size() ) );
+  if( pos_ < long( buffer_->size() ) )
     ( *buffer_ )[pos_] = vert;
   else
     buffer_->push_back( vert );
   ++pos_;
-  int capacity_after = buffer_->capacity();
+  long capacity_after = buffer_->capacity();
   log_capacity_change( log_capacity_changes_, capacity_before,
                        capacity_after );
 }
 
-void Emitter::set_position( int new_pos ) {
+void Emitter::set_position( long new_pos ) {
   pos_ = new_pos;
-  DCHECK( pos_ <= int( buffer_->size() ) );
+  DCHECK( pos_ <= long( buffer_->size() ) );
 }
 
 void Emitter::emit( span<GenericVertex const> vertices ) {
   if( vertices.empty() ) return;
-  int       capacity_before = buffer_->capacity();
-  int       current_size    = buffer_->size();
-  int const needed_size     = pos_ + vertices.size();
+  long       capacity_before = buffer_->capacity();
+  long       current_size    = buffer_->size();
+  long const needed_size     = pos_ + vertices.size();
   if( current_size <= needed_size )
     buffer_->resize( needed_size );
-  int capacity_after = buffer_->capacity();
+  long capacity_after = buffer_->capacity();
   log_capacity_change( log_capacity_changes_, capacity_before,
                        capacity_after );
-  DCHECK( int( buffer_->size() ) >= needed_size );
+  DCHECK( long( buffer_->size() ) >= needed_size );
   copy( vertices.begin(), vertices.end(),
         buffer_->begin() + pos_ );
   pos_ += vertices.size();
